+++ tests/bottleneck.rs
@@ -100,6 +100,7 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
                 approximation_bound: Time(0.0),
                 max_pending_duration: Time(f64::INFINITY),
                 spillback: false,
+                backward_wave_speed: None,
                 algorithm_type: Default::default(),
             }),
         },